Use Int for groupCount

+++ src/main/scala/adventofcode/common/IterableImplicits.scala
@@ -1,7 +1,7 @@
 package adventofcode.common
 
 implicit class IterableImplicits[A](it: Iterable[A]) {
-  def groupCount[K](f: A => K): Map[K, Long] = {
-    it.groupMapReduce(f)(_ => 1L)(_ + _)
+  def groupCount[K](f: A => K): Map[K, Int] = {
+    it.groupMapReduce(f)(_ => 1)(_ + _)
   }
 }
